backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.29.0

# Database
sqlalchemy>=2.0.30
alembic>=1.13.1
psycopg2-binary>=2.9.10

# Settings & Validation
pydantic[email]>=2.9.0
pydantic-settings>=2.5.0

# Auth & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# File Import
openpyxl>=3.1.0

# Image Processing
Pillow>=10.0.0

# Rate Limiting
slowapi>=0.1.9
